major cleanup on aisle livias code
