+++ client/src/data/fishtypes.ts
@@ -1,0 +1,20 @@
+
+
+import type { FishTemplate } from "../types/fish";
+
+//temp fish data for testing inventory
+ export  const fishTypes: FishTemplate[] = [
+  { name: "Salmon", rarity: "common" },
+  { name: "Tuna", rarity: "uncommon" },
+  { name: "Bass", rarity: "common" },
+  { name: "Pufferfish", rarity: "rare" },
+  { name: "Anglerfish", rarity: "rare" },
+  { name: "Golden Koi", rarity: "legendary" },
+  { name: "Dragonfish", rarity: "mythical" },
+  { name: "livia fish", rarity: "the one that got away", length: 170, weight: "how rude to ask"},
+  { name: "axel fish", rarity: "the one that got away", length: 177, weight: "how rude to ask"},
+  { name: "marcus fish", rarity: "the one that got away", length: 180, weight: "how rude to ask"},
+  { name: "jake fish", rarity: "the one that got away", length: 187, weight: "how rude to ask"},
+  { name: "josh fish", rarity: "the one that got away", length: 180, weight: "how rude to ask"},
+
+];
