+++ client/src/data/lakeDim.ts
@@ -1,0 +1,2 @@
+export const LakeWidth = 800;
+export const LakeHeight = 500;
