+++ client/src/utils/catchFish.ts
@@ -1,0 +1,7 @@
+import type { Fishy} from "../types/fish";
+import {setState } from 
+
+
+const catchFish = (fish: Fishy) => {
+    setState("caught");
+  }
