+++ client/src/utils/fishCreate.ts
@@ -1,0 +1,116 @@
+import { fishTypes } from "../data/fishtypes";
+import type { Fishy } from "../types/fish";
+import {LakeHeight,LakeWidth} from '../data/lakeDim'
+
+export  const createFish = (): Fishy => {
+    const roll = Math.random();
+    const side=Math.floor(Math.random() * 4);
+
+    let fish;
+
+// spawn outside the lake so it has to swim in
+    let x=0;
+    let y=0;
+    
+    if (side === 0) {
+      // left
+      x = -20;
+      y = Math.random() * LakeHeight;
+    } else if (side === 1) {
+      // right
+      x = LakeWidth + 20;
+      y = Math.random() * LakeHeight;
+    } else if (side === 2) {
+      // top
+      x = Math.random() * LakeWidth;
+      y = -20;
+    } else {
+      // bottom
+      x = Math.random() * LakeWidth;
+      y = LakeHeight + 20;
+    }
+
+    const centerX = LakeWidth / 2;
+    const centerY = LakeHeight / 2;
+
+    const vx = (centerX - x) * 0.001;
+    const vy = (centerY - y) * 0.001;
+
+
+
+    if (roll < 0.6) {
+      const commonFish = fishTypes.filter(f => f.rarity === "common");
+      fish = commonFish[Math.floor(Math.random() * commonFish.length)];
+    } else if (roll < 0.75) { 
+      const uncommonFish = fishTypes.filter(f => f.rarity === "uncommon");
+      fish = uncommonFish[Math.floor(Math.random() * uncommonFish.length)];
+    } else if (roll < 0.87) {
+      const rareFish = fishTypes.filter(f => f.rarity === "rare");
+      fish = rareFish[Math.floor(Math.random() * rareFish.length)];
+    } else if (roll < 0.9) {
+      const legendaryFish = fishTypes.filter(f => f.rarity === "legendary");
+      fish = legendaryFish[Math.floor(Math.random() * legendaryFish.length)];
+    } else if (roll < 0.99) {
+      const mythicalFish = fishTypes.filter(f => f.rarity === "mythical");
+      fish = mythicalFish[Math.floor(Math.random() * mythicalFish.length)];
+    } else {
+      const gotAwayFish = fishTypes.filter(f => f.rarity === "the one that got away");
+      fish = gotAwayFish[Math.floor(Math.random() * gotAwayFish.length)];
+    } 
+
+    // use fixed stats
+    if (fish.rarity === "the one that got away") {
+      return {
+        id: crypto.randomUUID(),
+        name: fish.name,
+        rarity: fish.rarity,
+        length: fish.length,   
+        weight: fish.weight ,
+
+      // spawn outside the lake so it has to swim in
+        x,y,
+
+        // initial movement
+        vx,
+        vy,
+        behavior: "swimming",
+        tapCount: 0,
+        tapCooldown: 0,
+        requiredTaps: Math.floor(Math.random() * 5) + 3, // random number of taps required to catch the fish 
+        tapflag:0, 
+        hX:0,
+        hY:0,
+        
+      };
+    }
+
+    const length=
+      fish.rarity === "common" ? Math.floor(Math.random() * 40) + 20 :
+      fish.rarity === "uncommon" ? Math.floor(Math.random() * 50) + 30 :
+      fish.rarity === "rare" ? Math.floor(Math.random() * 60) + 40 :
+      fish.rarity === "legendary" ? Math.floor(Math.random() * 80) + 50 :
+      Math.floor(Math.random() * 80) + 60;
+
+    const weight = Math.round(length * (Math.random() * 0.5 + 0.02)); // weight is based on length with some randomness 
+
+
+  return {
+    id: crypto.randomUUID(),
+    name: fish.name,
+    rarity: fish.rarity as Fishy["rarity"],
+    length: length,
+    weight: weight,
+
+    x,
+    y,
+    vx,
+    vy,
+    behavior: "swimming",
+    tapCount: 0,
+    tapCooldown: 0,
+    requiredTaps: Math.floor(Math.random() * 5) + 3, // random number of taps required to catch the fish
+    tapflag:0,
+     hX:0,
+    hY:0,
+  };
+};
